Changed num_bins from float to int
Changed CG_ENG_MPIC_WAKE_CPM from float to int

Changed CG_ENG_MPIC_WAKE_CPM from float to int

Changed LAMP_TIME from float to int

+++ mi/dataset/parser/cg_cpm_eng_cpm.py
@@ -1151,7 +1151,7 @@ PARAM_REGEX_RULES_AND_VALUES = [
       [CgCpmEngCpmParserDataParticleKey.CG_ENG_MPIC_HBEAT_THRESHOLD, int, None],
      ), ExpectedInstancesEnum.ONE),
     (MPIC_WAKE_CPM_REGEX,
-     ([CgCpmEngCpmParserDataParticleKey.CG_ENG_MPIC_WAKE_CPM, float, None],
+     ([CgCpmEngCpmParserDataParticleKey.CG_ENG_MPIC_WAKE_CPM, int, None],
       [CgCpmEngCpmParserDataParticleKey.CG_ENG_MPIC_WPC, int, None],
      ), ExpectedInstancesEnum.ONE),
     (MPIC_STC_EFLAG2_REGEX,
